Inner class ParseForMovieTralierAndReviews for fetching MovieTrailer and MovieReview
It's time to store MovieTrailer and MovieReview into database.
