+++ app/src/main/java/nanodegree/udacity/leon/udacitypopularmovies/helper/GeneralHelper.java
@@ -4,8 +4,17 @@ import android.content.Context;
 import android.content.SharedPreferences;
 import android.content.res.Configuration;
 import android.preference.PreferenceManager;
+import android.util.Log;
+
+import java.io.BufferedReader;
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.InputStreamReader;
+import java.net.HttpURLConnection;
+import java.net.URL;
 
 public class GeneralHelper {
+    private static final String LOG_TAG = GeneralHelper.class.getSimpleName();
 
     public static boolean isTablet(Context context) {
         return (context.getResources().getConfiguration().screenLayout
@@ -29,5 +38,57 @@ public class GeneralHelper {
         SharedPreferences sharedPreferences = PreferenceManager
                 .getDefaultSharedPreferences(context);
         return sharedPreferences.getInt(key, defaultValue);
+    }
+
+    /**
+     * Method is to get all the json data from the input URL as String.
+     * <improvement>: there should be URL check methods
+     *
+     * @param defaultUrl
+     * @return
+     */
+    public static String getAllJsonDataAsStringFromAPI(URL defaultUrl) {
+        HttpURLConnection urlConnection = null;
+        BufferedReader reader = null;
+        String moviesJsonStr = null;
+        try {
+            urlConnection = (HttpURLConnection) defaultUrl.openConnection();
+            urlConnection.setRequestMethod("GET");
+            urlConnection.connect();
+
+            InputStream inputStream = urlConnection.getInputStream();
+//                Log.v(LOG_TAG, "inputStream - getAllJsonDataAsStringFromAPI(): " + inputStream.toString());
+            StringBuffer buffer = new StringBuffer();
+            if (inputStream == null) {
+                return null;
+            }
+
+            reader = new BufferedReader(new InputStreamReader(inputStream));
+
+            String line;
+            while ((line = reader.readLine()) != null) {
+                buffer.append(line + "\n");
+            }
+
+            if (buffer.length() == 0) {
+                return null;
+            }
+            moviesJsonStr = buffer.toString();
+
+        } catch (IOException ex) {
+            ex.printStackTrace();
+        } finally {
+            if (urlConnection != null) urlConnection.disconnect();
+            if (reader != null) {
+                try {
+                    reader.close();
+                } catch (IOException e) {
+                    Log.e(LOG_TAG, "Error closing stream", e);
+                }
+            }
+        }
+
+//            Log.v(LOG_TAG, "moviesJsonStr - getAllJsonDataAsStringFromAPI(): " + moviesJsonStr);
+        return moviesJsonStr;
     }
 }
